requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
redis==5.0.1
python-socketio==5.10.0
python-engineio==4.8.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
numpy==1.26.2
scipy==1.11.4
python-multipart==0.0.6
aiofiles==23.2.1
